tests/requirements.txt
pytest
testcontainers
requests
requests[socks]
dnspython